修复V1.10因为版本号读取错误的各种奇葩报错，更新版本号判断规则，更新剩余天数计算（显示2位小数），更新导出激活码的命令

+++ src/me/zhanshi123/VipSystem/caches/PlaceholderCache.java
@@ -27,7 +27,7 @@ public class PlaceholderCache extends Cache{
 		String month=date.get(1);
 		String day=date.get(2);
 		String left=date.get(3);
-		long days=0;
+		float days=0;
 		SimpleDateFormat format=new SimpleDateFormat("yyyy-MM-dd");
 		try {
 			Date now=format.parse(year+"-"+month+"-"+day);
